requirements.txt
anyio==4.9.0
asgiref==3.8.1
babel==2.17.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-database-url==3.0.0
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.8.1
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-filter==25.1
django-money==3.5.4
django-redis==6.0.0
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docker==7.1.0
drf-yasg==1.21.10
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==25.0
psycopg2-binary==2.9.10
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
watchfiles==1.0.5
whitenoise==6.9.0
